Pages: first pass at moving Save/Clone buttons
Due to the layout of the page, and for consistency with other panels in the Presentation area, the submit button resides outside the main form. This is not an issue in HTML 5 as it allows the 'form' attribute to associate input tags with a master form elsewhere on the page.

To facilitate this, a data attribute has been added to any clone actions that lie outside their encompassing form. This tells the javascript to which form to attach the copy input tags that then permit the clone process to complete.